requirements.txt
screenpy
pytest-html
pytest-bdd
pytest-json-report

